+++ R/gg_boxplot.R
@@ -136,6 +136,13 @@ gg_boxplot <- function(data,
   }
   
   if (stat == "boxplot" & !is.numeric(y_var_vctr)) stop("Please use a numeric y variable for a boxplot when stat = 'boxplot'")
+  
+  if(is.logical(x_var_vctr)) {
+    data <- data %>% 
+      dplyr::mutate(dplyr::across(!!x_var, ~factor(., levels = c("TRUE", "FALSE"))))
+    
+    x_var_vctr <- dplyr::pull(data, !!x_var)
+  }
 
   if (is.null(x_title)) x_title <- snakecase::to_sentence_case(rlang::as_name(x_var))
   if (is.null(y_title)) y_title <- snakecase::to_sentence_case(rlang::as_name(y_var))
@@ -242,7 +249,7 @@ gg_boxplot <- function(data,
         labels = x_labels
       )
   }
-  else if (is.character(x_var_vctr) | is.factor(x_var_vctr) | is.logical(x_var_vctr)){
+  else if (is.character(x_var_vctr) | is.factor(x_var_vctr)){
     if(is.null(x_expand)) x_expand <- waiver()
     if(is.null(x_labels)) x_labels <- function(x) snakecase::to_sentence_case(x)
     
@@ -459,6 +466,19 @@ gg_boxplot_col <- function(data,
   if (stat == "boxplot" & !is.numeric(y_var_vctr)) stop("Please use a numeric y variable for a boxplot when stat = 'boxplot'")
   if (is.numeric(col_var_vctr)) stop("Please use a categorical colour variable for a boxplot")
   
+  if(is.logical(x_var_vctr)) {
+    data <- data %>% 
+      dplyr::mutate(dplyr::across(!!x_var, ~factor(., levels = c("TRUE", "FALSE"))))
+    
+    x_var_vctr <- dplyr::pull(data, !!x_var)
+  }
+  if(is.logical(col_var_vctr)) {
+    data <- data %>% 
+      dplyr::mutate(dplyr::across(!!col_var, ~factor(., levels = c("TRUE", "FALSE"))))
+    
+    col_var_vctr <- dplyr::pull(data, !!col_var)
+  }
+  
   if (is.null(x_title)) x_title <- snakecase::to_sentence_case(rlang::as_name(x_var))
   if (is.null(y_title)) y_title <- snakecase::to_sentence_case(rlang::as_name(y_var))
   if (is.null(col_title)) col_title <- snakecase::to_sentence_case(rlang::as_name(col_var))
@@ -579,7 +599,7 @@ gg_boxplot_col <- function(data,
         labels = x_labels
       )
   }
-  else if (is.character(x_var_vctr) | is.factor(x_var_vctr) | is.logical(x_var_vctr)){
+  else if (is.character(x_var_vctr) | is.factor(x_var_vctr)){
     if(is.null(x_expand)) x_expand <- waiver()
     if(is.null(x_labels)) x_labels <- function(x) snakecase::to_sentence_case(x)
     
@@ -689,7 +709,7 @@ gg_boxplot_col <- function(data,
 #' @param x_expand A vector of range expansion constants used to add padding to the x scale, as per the ggplot2 expand argument in ggplot2 scales functions. 
 #' @param x_labels A function or vector to modify x scale labels, as per the ggplot2 labels argument in ggplot2 scales functions. If NULL, categorical variable labels are converted to sentence case. Use ggplot2::waiver() to keep x labels untransformed.
 #' @param x_na TRUE or FALSE of whether to include x_var NA values. Defaults to TRUE.
-#' @param x_pretty_n For a numeric or date x variable, the desired number of intervals on the x scale, as calculated by the pretty algorithm. Defaults to 6. 
+#' @param x_pretty_n For a numeric or date x variable, the desired number of intervals on the x scale, as calculated by the pretty algorithm. Defaults to 3. 
 #' @param x_rev For a categorical x variable, TRUE or FALSE of whether the x variable variable is reversed. Defaults to FALSE.
 #' @param x_title X scale title string. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.
 #' @param x_title_wrap Number of characters to wrap the x title to. Defaults to 50. 
@@ -699,7 +719,7 @@ gg_boxplot_col <- function(data,
 #' @param y_expand A vector of range expansion constants used to add padding to the y scale, as per the ggplot2 expand argument in ggplot2 scales functions. 
 #' @param y_labels A function or vector to modify y scale labels, as per the ggplot2 labels argument in ggplot2 scales functions. If NULL, categorical variable labels are converted to sentence case. Use ggplot2::waiver() to keep y labels untransformed.
 #' @param y_na TRUE or FALSE of whether to include y_var NA values. Defaults to TRUE.
-#' @param y_pretty_n For a numeric or date x variable, the desired number of intervals on the x scale, as calculated by the pretty algorithm. Defaults to 5. 
+#' @param y_pretty_n For a numeric or date x variable, the desired number of intervals on the x scale, as calculated by the pretty algorithm. Defaults to 4. 
 #' @param y_title y scale title string. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.
 #' @param y_title_wrap Number of characters to wrap the y title to. Defaults to 50. 
 #' @param y_trans For a numeric y variable, a string specifying a transformation for the y scale, such as "log10" or "sqrt". Defaults to "identity".
@@ -742,7 +762,7 @@ gg_boxplot_facet <- function(data,
                              x_expand = NULL,
                              x_labels = NULL,
                              x_na = TRUE,
-                             x_pretty_n = 6,
+                             x_pretty_n = 3,
                              x_rev = FALSE,
                              x_title = NULL,
                              x_title_wrap = 50,
@@ -752,7 +772,7 @@ gg_boxplot_facet <- function(data,
                              y_expand = NULL,
                              y_labels = waiver(),
                              y_na = TRUE,
-                             y_pretty_n = 5,
+                             y_pretty_n = 4,
                              y_title = NULL,
                              y_title_wrap = 50,
                              y_trans = "identity",
@@ -803,6 +823,19 @@ gg_boxplot_facet <- function(data,
     
     if (stat == "boxplot" & !is.numeric(y_var_vctr)) stop("Please use a numeric y variable for a boxplot when stat = 'boxplot'")
     if (is.numeric(facet_var_vctr)) stop("Please use a categorical facet variable for a boxplot")
+    
+    if(is.logical(x_var_vctr)) {
+      data <- data %>% 
+        dplyr::mutate(dplyr::across(!!x_var, ~factor(., levels = c("TRUE", "FALSE"))))
+      
+      x_var_vctr <- dplyr::pull(data, !!x_var)
+    }
+    if(is.logical(facet_var_vctr)) {
+      data <- data %>% 
+        dplyr::mutate(dplyr::across(!!facet_var, ~factor(., levels = c("TRUE", "FALSE"))))
+      
+      facet_var_vctr <- dplyr::pull(data, !!facet_var)
+    }
     
     if (is.null(x_title)) x_title <- snakecase::to_sentence_case(rlang::as_name(x_var))
     if (is.null(y_title)) y_title <- snakecase::to_sentence_case(rlang::as_name(y_var))
@@ -919,7 +952,7 @@ gg_boxplot_facet <- function(data,
             labels = x_labels
           )
       }
-      else if (is.character(x_var_vctr) | is.factor(x_var_vctr) | is.logical(x_var_vctr)){
+      else if (is.character(x_var_vctr) | is.factor(x_var_vctr)){
         if(is.null(x_expand)) x_expand <- waiver()
         if(is.null(x_labels)) x_labels <- function(x) snakecase::to_sentence_case(x)
         
@@ -1004,7 +1037,7 @@ gg_boxplot_facet <- function(data,
 #' @param x_expand A vector of range expansion constants used to add padding to the x scale, as per the ggplot2 expand argument in ggplot2 scales functions. 
 #' @param x_labels A function or vector to modify x scale labels, as per the ggplot2 labels argument in ggplot2 scales functions. If NULL, categorical variable labels are converted to sentence case. Use ggplot2::waiver() to keep x labels untransformed.
 #' @param x_na TRUE or FALSE of whether to include x_var NA values. Defaults to TRUE.
-#' @param x_pretty_n For a numeric or date x variable, the desired number of intervals on the x scale, as calculated by the pretty algorithm. Defaults to 6. 
+#' @param x_pretty_n For a numeric or date x variable, the desired number of intervals on the x scale, as calculated by the pretty algorithm. Defaults to 3. 
 #' @param x_rev For a categorical x variable, TRUE or FALSE of whether the x variable variable is reversed. Defaults to FALSE.
 #' @param x_title X scale title string. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.
 #' @param x_title_wrap Number of characters to wrap the x title to. Defaults to 50. 
@@ -1014,7 +1047,7 @@ gg_boxplot_facet <- function(data,
 #' @param y_expand A vector of range expansion constants used to add padding to the y scale, as per the ggplot2 expand argument in ggplot2 scales functions. 
 #' @param y_labels A function or vector to modify y scale labels, as per the ggplot2 labels argument in ggplot2 scales functions. If NULL, categorical variable labels are converted to sentence case. Use ggplot2::waiver() to keep y labels untransformed.
 #' @param y_na TRUE or FALSE of whether to include y_var NA values. Defaults to TRUE.
-#' @param y_pretty_n For a numeric or date x variable, the desired number of intervals on the x scale, as calculated by the pretty algorithm. Defaults to 5. 
+#' @param y_pretty_n For a numeric or date x variable, the desired number of intervals on the x scale, as calculated by the pretty algorithm. Defaults to 4. 
 #' @param y_title y scale title string. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.
 #' @param y_title_wrap Number of characters to wrap the y title to. Defaults to 50. 
 #' @param y_trans For a numeric y variable, a string specifying a transformation for the y scale, such as "log10" or "sqrt". Defaults to "identity".
@@ -1075,7 +1108,7 @@ gg_boxplot_col_facet <- function(data,
                                  x_balance = FALSE,
                                  x_labels = NULL,
                                  x_na = TRUE,
-                                 x_pretty_n = 6,
+                                 x_pretty_n = 3,
                                  x_expand = NULL,
                                  x_rev = FALSE,
                                  x_title = NULL,
@@ -1086,7 +1119,7 @@ gg_boxplot_col_facet <- function(data,
                                  y_expand = NULL,
                                  y_labels = waiver(),
                                  y_na = TRUE,
-                                 y_pretty_n = 5,
+                                 y_pretty_n = 4,
                                  y_title = NULL,
                                  y_title_wrap = 50,
                                  y_trans = "identity",
@@ -1150,6 +1183,25 @@ gg_boxplot_col_facet <- function(data,
   if (stat == "boxplot" & !is.numeric(y_var_vctr)) stop("Please use a numeric y variable for a boxplot when stat = 'boxplot'")
   if (is.numeric(col_var_vctr)) stop("Please use a categorical colour variable for a boxplot")
   if (is.numeric(facet_var_vctr)) stop("Please use a categorical facet variable for a boxplot")
+  
+  if(is.logical(x_var_vctr)) {
+    data <- data %>% 
+      dplyr::mutate(dplyr::across(!!x_var, ~factor(., levels = c("TRUE", "FALSE"))))
+    
+    x_var_vctr <- dplyr::pull(data, !!x_var)
+  }
+  if(is.logical(col_var_vctr)) {
+    data <- data %>% 
+      dplyr::mutate(dplyr::across(!!col_var, ~factor(., levels = c("TRUE", "FALSE"))))
+    
+    col_var_vctr <- dplyr::pull(data, !!col_var)
+  }
+  if(is.logical(facet_var_vctr)) {
+    data <- data %>% 
+      dplyr::mutate(dplyr::across(!!facet_var, ~factor(., levels = c("TRUE", "FALSE"))))
+    
+    facet_var_vctr <- dplyr::pull(data, !!facet_var)
+  }
   
   if (is.null(x_title)) x_title <- snakecase::to_sentence_case(rlang::as_name(x_var))
   if (is.null(y_title)) y_title <- snakecase::to_sentence_case(rlang::as_name(y_var))
@@ -1273,7 +1325,7 @@ gg_boxplot_col_facet <- function(data,
           labels = x_labels
         )
     }
-    else if (is.character(x_var_vctr) | is.factor(x_var_vctr) | is.logical(x_var_vctr)){
+    else if (is.character(x_var_vctr) | is.factor(x_var_vctr)){
       if(is.null(x_expand)) x_expand <- waiver()
       if(is.null(x_labels)) x_labels <- function(x) snakecase::to_sentence_case(x)